app/requirements.txt
fastapi==0.115.12
mlflow-skinny==2.22.1
numpy==2.3.0
pandas==2.2.3
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
pydantic==2.11.5
python-json-logger==3.3.0
requests==2.32.4
scikit-learn==1.7.0